Separated logich from component in custom hook and types in types file.
