+++ src/components/Form/Select/Types.ts
@@ -1,0 +1,4 @@
+export interface Option {
+  label: string;
+  value: string;
+}
