+++ src/containers/Home/Form/index.tsx
@@ -3,6 +3,7 @@ import Select from "../../../components/Form/Select";
 import Search from "./Search";
 
 const options = [
+  { label: "All Regions", value: "all" },
   { label: "Africa", value: "africa" },
   { label: "America", value: "america" },
   { label: "Asia", value: "asia" },
